For better or worse increased solar radius to match wikipedia
I also changed the constant in the Hosek sky model, without being sure
if this is OK. Looking at the disk directly, the perceived difference is very
small, but the total sun light energy increased noticeably.
Effectively, scene lit by the sky will appear less blue, after
adjusting exposure. Overall, I like the change, but don't know if I
decreased physical accuracy.

Also forcing the sun/sky colors to be >= 0 now. This was a bug that
could happen for low sun elevation. Not sure whether this is the
correct way to handle this, or if there is a problem in my Spectral ->
XYZ -> sRGB conversion.