Removed (almost) all exceptions - expect some hick-ups
I settled on the most simple solution I could think of: returning
nullptr and logging where the error happens. I am not completely happy
with this for two reasons:

1. You cannot log from base module, so json parsing has an error out
parameter that can be logged from the calling site.

2. Some useful error messages are no split into two log entries
because not all information is always available. For example the
filename is often not available in the reader when the error happens.

I'm still committed to "exception-less" but maybe I can find a better compromise.

cxxopts still uses exceptions for error reporting unfortunately, so I
have to think of something there as well.