Changed entity topology to from ptr to id
This makes the Entity and Prop objects slightly smaller.
Usage is now slightly awkward because you have to pass the entity
array around. I don't think the change in itself is worth
much (actually the interface got a bit more clunky), but it is now
a little bit easier to separate out the topology from the Entity
class. Doubtful if that is ever going to happen, or indeed, what the
benefits would be, but I was hyped by all the DoD talk.