Re-introduced Stream_ptr, but differently
Usage is almost as before, except that you still have to go through
file::System to get a stringstream. Filesystem doesn't use heap
allocation with those streams. Main usage is to close the streams when
they go out of scope, which might be useful if we ever need multiple
streams at the same time.